+++ tests/test_order_Nematic.py
@@ -1,23 +1,18 @@
 import numpy as np
 import numpy.testing as npt
 import pytest
-import rowan
 
 import freud
 
 
 class TestNematicOrder:
     def test_perfect(self):
-        """Test perfectly aligned systems with different molecular axes"""
+        """Test different perfectly aligned systems."""
         N = 10000
-        axes = np.zeros(shape=(N, 3), dtype=np.float32)
-        angles = np.zeros(shape=N, dtype=np.float32)
-        axes[:, 0] = 1.0
-        orientations = rowan.from_axis_angle(axes, angles)
+        u = [1, 0, 0]
+        orientations = np.repeat(np.expand_dims(u, axis=0), repeats=N, axis=0)
 
-        # Test for parallel to molecular axis
-        u = np.array([1, 0, 0])
-        op_parallel = freud.order.Nematic(u)
+        op_parallel = freud.order.Nematic()
 
         # Test access
         with pytest.raises(AttributeError):
@@ -41,12 +36,13 @@ class TestNematicOrder:
         npt.assert_equal(op_parallel.director, u)
         npt.assert_equal(op_parallel.nematic_tensor, np.diag([1, -0.5, -0.5]))
         npt.assert_equal(
-            op_parallel.nematic_tensor, np.mean(op_parallel.particle_tensor, axis=0)
+            op_parallel.nematic_tensor,
+            np.mean(op_parallel.particle_tensor, axis=0),
         )
 
-        # Test for perpendicular to molecular axis
-        u = np.array([0, 1, 0])
-        op_perp = freud.order.Nematic(u)
+        u = [0, 1, 0]
+        orientations = np.repeat(np.expand_dims(u, axis=0), repeats=N, axis=0)
+        op_perp = freud.order.Nematic()
         op_perp.compute(orientations)
 
         assert op_perp.order == 1
@@ -60,12 +56,14 @@ class TestNematicOrder:
         """
         N = 10000
         np.random.seed(0)
+        u = [1, 0, 0]
 
-        # Generate orientations close to the identity quaternion
-        orientations = rowan.interpolate.slerp([1, 0, 0, 0], rowan.random.rand(N), 0.1)
+        # Generate orientations close to the u
+        orientations = np.random.normal(
+            np.repeat(np.expand_dims(u, axis=0), repeats=N, axis=0), 0.1
+        )
 
-        u = np.array([1, 0, 0])
-        op = freud.order.Nematic(u)
+        op = freud.order.Nematic()
         op.compute(orientations)
 
         npt.assert_allclose(op.order, 1, atol=1e-1)
@@ -78,13 +76,15 @@ class TestNematicOrder:
         assert not np.all(op.nematic_tensor == np.diag([1, -0.5, -0.5]))
 
         u = np.array([0, 1, 0])
-        op_perp = freud.order.Nematic(u)
+        orientations = np.random.normal(
+            np.repeat(np.expand_dims(u, axis=0), repeats=N, axis=0), 0.1
+        )
+        op_perp = freud.order.Nematic()
         op_perp.compute(orientations)
 
         npt.assert_allclose(op_perp.order, 1, atol=1e-1)
         assert op_perp.order != 1
 
-        # The director can be inverted so we compare absolute values
         npt.assert_allclose(np.abs(op_perp.director), u, atol=1e-1)
         assert not np.all(op_perp.director == u)
 
@@ -92,6 +92,5 @@ class TestNematicOrder:
         assert not np.all(op_perp.nematic_tensor == np.diag([-0.5, 1, -0.5]))
 
     def test_repr(self):
-        u = np.array([1, 0, 0])
-        op = freud.order.Nematic(u)
+        op = freud.order.Nematic()
         assert str(op) == str(eval(repr(op)))